0.6 feature dev (#25)
see changelog for 0.6

+++ src/main/java/game/map/editor/geometry/primitive/StairGenerator.java
@@ -177,7 +177,7 @@ public class StairGenerator extends ShapeGenerator
 		Vertex[] topL = new Vertex[N];
 
 		int R = Math.round(width / 2.0f);
-		float uvScale = UV_SCALE / width;
+		float uvScale = (float) UV_SCALE / width;
 		int vOffset = (UV_SCALE * 3) / 2;
 
 		for (int i = 0; i < N; i++) {
